除算メソッドのテスト

+++ src/main/java/junit/tutorial/Calculator.java
@@ -8,7 +8,7 @@ public class Calculator {
         return x * y;
     }
 
-    public int divide(int x, int y){
-        return x / y;
+    public float divide(int x, int y){
+        return (float) x / (float) y;
     }
 }
